Intern comments on declare module exports statements
Summary:
Adding comment attachment for the handful of remaining declare statements for completeness. This change adds comment attachment for declare module exports statements.

Comments before the `declare`, `module`, `.`, `exports`, or `:` are attached as leading comments on the declare module exports statement.

After the semicolon is a type, which will have leading and trailing comments already attached to it.

Finally, the declare module exports statement optionally ends in a semicolon. If a semicolon is supplied then comments after the semicolon are attached as trailing the entire statement, following normal trailing comment attachment rules. If a semicolon is not provided, trailing comments may have to be remved from the type annotation using the trailing comments remover.

Full comment attachment is shown below:
```
/* L exports */ declare /* L exports */ module /* L exports */ . /* L exports */ exports /* L exports */ : /* L any */ any /* T any */ ; /* T exports */
```

Note that this involved creating a new record type in the AST for declare module exports, so they can hold both the type annotation and attached comments.

Reviewed By: mroch

Differential Revision: D20822379

fbshipit-source-id: 4fc179b4dac3308f3d8c9322eacc2247fcfda6af