Intern leading comments in object type properties
Summary:
Object type properties can have leading comments before a number of optional keywords that can be used when object type is for a class. Comments that directly precede the `static`, `proto`, `get`, and `set` keywords will be attached as leading the property node. This applies both to regular object type property nodes, as well as call property, indexer, and internal slot nodes.

```
{
  /* L prop */ static /* L id */ prop: any,
  /* L prop */ proto /* L id */ prop: any,
  /* L prop */ static /* L prop */ get /* L id */ getter(): void,
  /* L prop */ static /* L prop */ set /* L id */ setter(any): void,
  /* L prop */ static /* L id */ method(): void,
  /* L call prop */ static /* L func */ (): void,
  /* L indexer */ static /* L indexer */ [string]: string,
  /* L islot */ static /* L islot */ [[prop]]: string,
}
```

We can accomplish this in the parser by passing down the leading comments when parsing object type properties, appending to the list of leading comments when a `static`, `proto`, `get`, or `set` token is encountered.

Reviewed By: gkz

Differential Revision: D20800788

fbshipit-source-id: 88abe0fcbeaf66eb97bab250a0a3179279bfaff6