Intern comments in enum declarations
Summary:
Adds comment attachment for enum declarations.

Comments can be attached to either the enum declaration node itself, or the inner enum body node. Comments that directly precede the `enum` keyword will be attached as leading the entire enum declaration.

Comments after the closing `}` of the enum's body are attached as trailing the enum body node. The regular rules for statement trailing comment attachment apply here - all trailing comments are attached if it is the last statement in its block or file, only trailing comments on that line are attached if followed by a line break, otherwise no trailing comments will be attached as they are all leading the next statement.

Comments around the enum's name will be attached to the name's identifier. If an explicit type is provided for the enum, then comments directly before the opening `{` of the enum body will be attached as leading the body node. We do not yet support comment attachment for comments in the middle of an explicit type (e.g. between the `of` and `string` keywords).

```
/* L enum */ enum /* L id */ Foo /* T id */ {} /* T enum body */
/* L enum */ enum /* L id */ Foo /* T id */ of string /* L enum body */ {} /* T enum body */
```

Comment attachment within enum bodies already works, since they simply consist of ids and literals separated by punctation. For example:

```
enum Foo {
  /* L id */ Bar /* T id */,
  /* L id */ Baz /* T id */,
}

enum Foo {
  /* L id */ Bar /* T id */ = /* L lit */ 10 /* T lit */,
}
```

Reviewed By: gkz

Differential Revision: D20821458

fbshipit-source-id: 685c4db8b9fac65d59e697058946ebd22cc6d39e