Intern comments in decorators
Summary:
Attaches comments to decorator nodes in the AST. Comments before the `@` starting a decorator will be attached as leading comments. The `@` is followed by an expression, which will have all adjacent comments greedily attached.

Since decorators always appear before either another decorator or a class/method, comments on  later lines may have been greedily attached as trailing the decorator's expression but should actually be attached as leading the next node. We can fix this by detecting this situation and using the trailing comments remover to remove comments that should not be trailing the expression.

If the decorator is followed by a line break then trailing comments on that line will be attached as trailing the decorator's expression. If the decorator is not followed by a line break then its expression will have no trailing comments as they will be attached as leading the next node.

```
/* L dec */ foo() /* T call */
/* L class */ class C {}

/* L dec */ foo() /* L dec */ bar /* L class */ class C {}
```

Reviewed By: mroch

Differential Revision: D20801714

fbshipit-source-id: e080e284f294271e573242b16f2f4174abe433d6