Intern comments in declare variable statements
Summary:
Attaches comments to `DeclareVariable` nodes in the AST. Comments directly preceding the `declare` or `var/let/const` keywords will be attached as leading comments.

`DeclareVariable` nodes are optionally closed by a trailing semicolon. If a semicolon is included then comments after that semicolon will be attached as trailing the `DeclareVariable` node. The same rules for statement comment attachment apply here.

```
/* L decl var */ declare /* L decl var */ var x; /* T decl var */
```

The inner id and type annotation will greedily attach comments to themselves, so full comment attachment for declare variable statements is shown below:

```
/* L decl var */ declare /* L decl var */ var /* L id */ x /* T id */ : /* L mixed */ mixed /* T mixed */ ; /* T decl var */
```

If declare variable statements are not followed by a semicolon then we must remove comments that were greedily attached to the last node, whether it be the variable name or type annotation. This can be done using the comments remover.

Reviewed By: gkz

Differential Revision: D20802945

fbshipit-source-id: 6c9f3f3fd4008130bc58494163aecbfd416520cc