Fix leading and trailing comment attachment on export default declarations
Summary:
Let's finish comment attachment on export default declarations. First of all, we should attach comments between the `export` and `default` keywords as leading the export default declaration. Second, we should handle trailing comment attachment. Trailing comment attachment for default exported functions, classes, and enums is already handled by the parsers for those nodes. The only case we need to handle is a default exported expression. In this case we must use the semicolon helper and the trailing comment remover to fix trailing comments on the expression if a semicolon is not provided.

```
/* L export */ export /* L export */ default 1; /* T export */
/* L export */ export /* L export */ default 1 /* T num */

```

Reviewed By: gkz

Differential Revision: D20820281

fbshipit-source-id: 6bfdd5002a776f9ccc06ea798d91a127bbcc5776