Fix trailing comment attachment in class bodies
Summary:
We currently greedily attach leading and trailing comments when parsing, however this leads to comments being attached to two nodes when a comment is ambiguously between two AST nodes. We must specify rules to disambiguate which node these comments should be attached to.

This change fixes trailing and duplicate comment attachment within class bodies - namely between class properties and methods.

## Properties

Properties are optionally followed by a semicolon. When followed by a semicolon, comments after the semicolon are attached to the property node.

If the property is the last property/method in a class (meaning it is followed by `T_EOF` or `T_RCURLY`) then all trailing comments should be attached. Else if the property is followed by a newline then comments on the same line as the property are attached as trailing. Otherwise the property must be on same line as the next element, so the property has no trailing comments since they are all attached as leading the next element.

```
class { prop = value; /* T prop */ }
class { prop = value; /* T id */ prop = value; }
class {
  prop = value; /* T prop */
  /* L id */ prop = value;
}
```

If a closing semicolon is not supplied then trailing comments currently could have been greedily attached to the last node in the property, which could be an initialization expression, type annotation, or the property name itself:

```
/* L id */ prop = 1 /* T num, L id */
/* T num, L id */ prop: any /* T any, L id */
/* T any, L id */ prop /* T id, L id */
/* T id, L id */ prop
```

The correct comment attachment with this change is the following:

```
/* L id */ prop = 1 /* T num */
/* L id */ prop: any /* T any */
/* L id */ prop /* T id */
/* L id */ prop
```

We can detect which case we are in during parsing and use the trailing comment remover to remove comments that were greedily attached but should actually be attached to the next method/property.

## Methods

Methods always end with function block bodies so trailing comment attachment is straightforward to implement. Like properties with semicolons, if a method is the last element in a class (next token is `T_EOF` or `T_RCURLY`) then all trailing comments are attached. Else if a method is followed by a newline then comments on the same line as the method are attached as trailing. Otherwise method must be on same line as next element, so the method has no trailing comments since they are all attached as leading the next method/property.

Reviewed By: mroch

Differential Revision: D20784842

fbshipit-source-id: 461a40f42e3cbaf2b8d97f36fcdab055ceb2c902