Intern comments in for loops
Summary:
It looks like the very last statement that still needs comments attached is the for statement (as well as for-in and for-of).

For all for statement types, comments before the `for` keyword and between the `for` and left paren will be attached as leading comments.

```
/* L for */ for /* L for */ (;;) {}
```

For-of statements can also have an optional `await` keyword included, so comments between the `for` and `await` keywords are also attached as leading comments.

These are the only comments attached to for statements - all other comments will be attached to children of the for node. Examples of complete comment attachment for for nodes are shown below:

```
/* L for */ for /* L for */ (/* L var decl */ let x = 1 /* T num */; /* L bool */ true /* T bool */; /* L id */ x /* T id */) /* L block {} /* T block */

/* L for */ for /* L for */ await /* L for */ (/* L id */ x /* T id */ of /* L arr */ [] /* T arr */) /* L block */ {} /* T block */

/* L for */ for /* L for */ (/* L id */ x /* T id */ in /* L obj */ {} /* T obj */) /* L block */ {} /* T block */
```

Reviewed By: mroch

Differential Revision: D20827097

fbshipit-source-id: f6a9fc06d9560b98fe6c4edc83e98b8e6fff0ecc