[tool] stop using --strip-root
Summary:
when fetching errors for add-comments and remove-comments, it used --strip-root so that the path shown in the GUI was relative, but then had to join it back together the rest of the time.  this makes it hard to compare two paths since they're not absolute.

I inverted it, so we get absolute paths in errors and relativize them for the UI

Reviewed By: pieterv

Differential Revision: D20886223

fbshipit-source-id: d3183dcf970cc9873f462c512329edae8ebbf119