Fix comment attachment for while and with statements
Summary:
Let's quickly fix while and with statement comment attachment to eliminate duplicates and make them consistent with handling of other statements.

First of all, comments between the `while` or `with` and the opening paren will be attached as leading comments, just like they are for if and for statements.

Also, comments after the closing paren should not be attached as trailing the entire statement. Instead, they will be attached as leading the body of the while or with statement. This also means that while/with statements will not have any trailing comments, as any trailing comments will be attached to the inner body statement node.

Reviewed By: gkz

Differential Revision: D20828735

fbshipit-source-id: 1999eb1c5d1b77b7497ce7c177302bfe8ab9b920