Fix comment attachment for if statements
Summary:
Fixes comment attachment for if statements, as comment attachment has slightly changed since this was first implemented.

If statements can have leading comments that appear before the `if` keyword and between the `if` and opening paren. Inside the parens, all comments will be attached to the test expression. Comments after the closing paren will be attached as leading the statement that is the consequent of the if statement (usually a block), instead of being attached to the if statement itself. This means that if statements do not have trailing comments, since trailing comments will be attached to the consequent statement (or alternate statement for if-else).

```
/* L if */ if /* L if */ (/* L bool */ true /* T bool */) /* L block */ {} /* T block */
```

This diff also explicitly interns comments in if alternate clauses, as the consequent clause is a statement so it only will keep trailing comments that are on the same line as the end of the consequent. This required creating a new `If.Alternate` node in the AST that contains the body of the alternate and attached comments. Comments directly before the `else` keyword will be attached as leading, and there are no trailing comments since they would be attached to the body of the alternate.

This allows us to have comment attachment in alternate clauses that respect newlines:

```
if (true) {
} /* L alternate */ else {
}

if (true) {
} /* T block */
/* L alternate */ else {
}
```

Note that estree does not have a great node for us to output these comments as attached to, so lets report them on the if node itself and combine them with the comments on the if node.

Reviewed By: mroch

Differential Revision: D20828170

fbshipit-source-id: d2b7870eee3671e4e62ce8e5e7806f029a4a4ac7