[react] react.runtime automatic
Summary:
React is adding a flag to their babel transform.

`runtime automatic` will auto-import `jsx` (the replacement for createElement` under a fresh new name.

`runtime classic` will continue transpiling jsx to `React.createElement` as it does today.

The transform guarantees that the name it uses for the auto-imported `jsx` will be fresh and not used elsewhere. On our side, we call directly into the `createElement` machinery without looking for the function name the transform would create. We do this instead of modeling the transform more directly so that we do not need to mimic babel's variable renaming process inside Flow. That implementation detail should not be something we need to track and update in Flow.

Reviewed By: mroch

Differential Revision: D20860562

fbshipit-source-id: 84fd8eaed0f1d1fd6a32567c186fbfc4da35eb95