Fix ambiguous adjacent comment attachment in arrow functions
Summary:
We are currently going through and fixing ambiguous comment attachment between adjacent nodes that could both have comments attached. The general rule is that trailing comments on the same line as the first node are attached as trailing that node, and all comments on later lines until the second node will be attached as leading the second node. If there is not a line break between the adjacent nodes, the first node will have no trailing comments as they will all be attached as leading the second node.

Let's fix these issues in arrow functions. Luckily this is very straightforward, as the only adjacent nodes that can both have comments attached are the type params and params list. We can fix these using the trailing comment remover for type params. We don't need to make any fixes for the return type/predicate as they are separated from the params by a semicolon, and both the params and return type are separated from the body with an arrow.

```
/* L tparams */ <T> /* L params */ () => {}

/* L tparams */ <T> /* T tparams */
  /* L params */ () /* T params */ => {}
```

Luckily we don't have to make any changes in the case where a single unparenthesized arugment is supplied, becuase type params are not supported with this syntax, and the identifier is separated from the return type with a colon and from the body with the arrow.

Reviewed By: mroch

Differential Revision: D20927153

fbshipit-source-id: 6de2608eb96604605eae6678ef85e67d24ccd49b