Fix ambiguous adjacent comment attachment for TypeArgs
Summary:
We are currently going through and fixing ambiguous comment attachment between adjacent nodes that could both have comments attached. The general rule is that trailing comments on the same line as the first node are attached as trailing that node, and all comments on later lines until the second node will be attached as leading the second node. If there is not a line break between the adjacent nodes, the first node will have no trailing comments as they will all be attached as leading the second node.

Let's fix these issues around `TypeArgs` nodes. `TypeArgs` can be used in three places - on generic types, superclasses in a class, and implemented interfaces in a class. In all three cases they are directly preceded by either an identifier or expression, which was previously greedily attaching all trailing comments, so `TypeArgs` could not have leading comments. Instead, let's make these follow the attachment rules outlined above by using the trailing comments remover on the preceding node.

Also creates a shared `Comment_attachment.trailing_and_remover` helper to create the trailing comments and trailing comments remover, depending on if there is currently a line break before the next node or if the next node is on the same line.

This means that comment attachment around type args is now the following:

```
// Same line
Foo /* L targs */ <T>

// Different lines
Foo /* T id */
/* L targs */ <T>
```

Reviewed By: mroch

Differential Revision: D20901633

fbshipit-source-id: 6225b43bc158b5d8e0c16aafe9307fcbd5970997