Intern comments and fix ambiguous adjacent comment attachment for new and call expressions
Summary:
We are currently going through and fixing ambiguous comment attachment between adjacent nodes that could both have comments attached. The general rule is that trailing comments on the same line as the first node are attached as trailing that node, and all comments on later lines until the second node will be attached as leading the second node. If there is not a line break between the adjacent nodes, the first node will have no trailing comments as they will all be attached as leading the second node.

This change fixes comment attachment within new and call expressions. For new and call expressions there are three adjacent nodes: the callee, the call type args, and the args list. The AST was updated to include a comments field on both  `CallTypeArgs` and `ArgsList`, both of which were previously simply lists of arguments. Now, leading and trailing comments can be attached to all three of these nodes.

```
// Call expressions
/* L id */ foo /* L targs */ <T> /* L arg list */ () /* T arg list */

/* L id */ foo /* T id */
/* L targs */ <T> /* T targs */
/* L arg list */ () /* T arg list */

// New expressions
/* L new */ new /* L id */ Foo /* L targs */ <T> /* L arg list */ () /* T arg list */

new /* L id */ Foo /* T id */
/* L targs */ <T> /* T targs */
/* L arg list */ () /* T arg list */
```

Note that this means that call expression will not have comments attached directly as the callee will get leading comments and the args list will get trailing comments, but still can have comments attached if the call expression is surrounded in parentheses. New expressions will have trailing attached to the args list as well, but can still have leading comments before the `new` keyword.

The arguments and type arguments are both optional for new expressions, so we must check after the callee and type args whether they are followed by another node before the end of the new expression in order to determine when to remove trailing comments.

Note that `ArgsList` do not have their own nodes in the estree output, so let's combine comments on `ArgsLists` with the comments on the call node  and report comments on the call node.

Reviewed By: mroch

Differential Revision: D20922515

fbshipit-source-id: 6bb9e67a3f3130796aec08052ce7a3b42b4a8368