Fix ambiguous adjacent comment attachment in type aliases
Summary:
We are currently going through and fixing ambiguous comment attachment between adjacent nodes that could both have comments attached. The general rule is that trailing comments on the same line as the first node are attached as trailing that node, and all comments on later lines until the second node will be attached as leading the second node. If there is not a line break between the adjacent nodes, the first node will have no trailing comments as they will all be attached as leading the second node.

Let's fix the instances of ambiguous adjacent comment attachment in type aliases and opaque type aliases. For type aliases the only case of ambiguity are comments between the type name and the optional type params. We can use the trailing comment remover to make it follow the rules of comment attached outline above.

```
type Foo /* L tparams */ <T> = any;

type Foo /* T id */
  /* L tparams */ <T> = any;
```

Reviewed By: mroch

Differential Revision: D20933673

fbshipit-source-id: 48f41db7fc27e767bc3884533e751af194f3850e