Fix error in printing of arrow function with function return type
Summary:
There is a bug in the `js_layout_generator` when printing arrow functions with function return types. When a function type is the return type of an arrow function, a parse error will occur as the arrow from the arrow function will be interpreted as part of the function type. To avoid this we must wrap the entire function type in parentheses if it is the return type of an arrow function.

```
// Errors
(): (x) => y => {}

// OK
(): ((x) => y) => {}
```

Note that this is only a syntax error when the function type has params. If the function type has an empty params list, or a params list that consists of just a rest param, then the function type does not need to be wrapped in parentheses.

```
// OK
(): () => y => {}
(): (...x) => y => {}
```

Reviewed By: gkz

Differential Revision: D20970397

fbshipit-source-id: b97095b928d88e3889b0a9df6e3807d11faefe47