Fix ambiguous adjacent comment attachment within tagged templates
Summary:
I found another case of ambiguous adjacent comment attachment when working on the `js_layout_generator`.

Comments between the tag and template literal of a tagged template are ambiguous as they could be attached to either node. In the current system they are accidentally attached to both nodes, which should not be happening and is causing problems with reprinting.

For resolution, the general rule is that trailing comments on the same line as the first node are attached as trailing that node, and all comments on later lines until the second node will be attached as leading the second node. If there is not a line break between the adjacent nodes, the first node will have no trailing comments as they will all be attached as leading the second node.

We can use the trailing comments remover to enforce this behavior.

```
tag /* L template lit */ `literal`

tag /* T id */
  /* L template lit */ `literal`
```

Reviewed By: pieterv

Differential Revision: D21006103

fbshipit-source-id: e78adb0294ce3166e71586e11cb4dc882968e5e8