[Enums] allow unions of enums to cast to their supertype
Summary:
Consider the following code:

```
enum B of boolean {
  A = true,
  B = false,
}

enum B2 of boolean {
  A = true,
  B = false,
}

declare var b : B;
(b : boolean); // Valid

declare var b2 : B2;
(b2 : boolean); // Valid

declare var bs: B | B2;
(bs: boolean); // Error

if (typeof bs === "boolean") {
(bs: boolean); // Valid
}
```

Enum casts behave slightly differently to other type casts, in that one can only cast an enum to their exact representation type. For this reason, the 3rd cast in this example program fails, since the lower bound is a union of enums. However, the 4th cast succeeds since the lower bound of this cast is a type variable with two bounds because of a refinement, so instead of performing a single union cast which fails, we perform two casts which succeed.

This is inconsistent and exposes an implementation detail that should not affect the user. We change the union cast behavior to make the 3rd cast in this example valid so that cast 3 and cast 4 have consistent behavior.

Reviewed By: gkz

Differential Revision: D20967950

fbshipit-source-id: e070970c69ce02ef6db47bc1cfe96bac41492889