Preserve newlines around comments in js_layout_generator
Summary:
When pretty printing comments in the `js_layout_generator` we want to preserve some of the newlines in the original file to keep comments separated and grouped as they were before. I think it would be sensible to follow the same rules as prettier.

When printing in pretty mode the following rules are followed, for inserting newlines and spaces between adjacent comments (or between a comment and the node it is attached to).

- Line comments are always terminated by a newline
- If two adjacent comments are on the same line then they will be printed on the same line, separated by a space
- If two adjacent comments are separated by exactly one or two newlines, then keep the same number of newlines when reprinting
- If two adjacent comments are separated by more than two newlines they will be reprinted as separated by two newlines

If we are not printing in pretty mode then no newlines will be preserved, except for the newline that is always present at the end of a line comment.

This was accomplished by rewriting the comment printing functions in `js_layout_generator`. For leading comments we need to know the loc of the next comment so that we can print the number of newlines separating the two. For trailing comments we need to know the loc and type of the previous comment, so we can print the number of newlines separating the two. Since line comments are always printed with a newline at the end, we must know whether the previous/next comment is a line comment so that we can accurately print the required extra newlines.

Reviewed By: pieterv

Differential Revision: D21007820

fbshipit-source-id: 04260e6d29665849e75a0a8318eec3d5fde06f89