Print trailing comments on statement after semicolon
Summary:
When printing trailing comments on statements, the trailing comments should be printed after the closing semicolon if one exists. I checked all statements to make sure the semicolon and trailing comments were being printed in the right order and fixed those that were not. This mostly required just flipping the calls to `with_semicolon` and `layout_node_with_comments_opt` in `js_layout_generator`.

The only non-trivial change that I made was to  printing variable declarations, which use a separate `variable_declaration` helper function. We can add an optional `semicolon` argument here filled with the correct pretty/regular semicolon, which will not be set for variable declarations inside for-in or for-of loops. Additionally this moves the semicolon inside the source location of the variable declaration, but I think this change makes sense since the semicolon is part of the variable declaration statement and its loc.

Reviewed By: pieterv

Differential Revision: D21009505

fbshipit-source-id: 063ea525d2b767c8daa416d3fbdde29c33ec7338