Do not print trailing comma on rest parameter
Summary:
We currently have a bug when printing function parameters in pretty mode. If function params break over multiple lines and contain a rest param, then a trailing comma will be printed after the rest param, which is a syntax error.

Instead, we can simply only print the trailing comma on the last param if it is not a rest param.

```
// Old printing, errors
function foo(
  ...x,
) {}

(
  ...x,
) => {}

// New printing, ok
function foo(
  ...x
) {}

(
  ...x
) => {}
```

Reviewed By: mroch, gkz

Differential Revision: D20973871

fbshipit-source-id: 962d70afb496acb296e7496f05ea98b77b9276e6