Print comments attached to literals in js_layout_generator
Summary: Literals were not having attached comments printed in the `js_layout_generator`. Let's fix this by passing the comments to `source_location_with_comments` when printing literals in `js_layout_generator`. This also requires saving the location of literals when creating the AST diff, as this location is necessary for printing comments.

Reviewed By: gkz

Differential Revision: D20998649

fbshipit-source-id: 101da81ce26cd1002c74a74a099eb144a5b8bbb0