Preserve empty lines when printing objects, object types, and classes
Summary:
When printing in pretty mode we want to preserve formatting in the same way as prettier, so that running prettier after will yield the original file with minimal changes. One such location where we want to  preserve some formatting is between adjacent properties or elements in objects (or object types) and classes, respectively.

The `list_with_newline` utility function will already handle inserting newlines between adjacent nodes if we pass it a list of (loc, comment bounds, node) pairs.

The result of this is that empty lines between adjacent properties/elements will be preserved when printing in pretty mode with the `js_layout_generator`. If there is more than one blank line separating adjacent properties/elements, it will be condensed to a single blank line.

This replaces the old pretty newline rules for object literals, where `pretty_hardlines` were inserted around properties that contained a function.

Reviewed By: pieterv

Differential Revision: D21032395

fbshipit-source-id: 4c353b17e4c3163cac7800c6004662be8919ad58