Include comments when preserving newlines between statements in js_layout_generator
Summary:
When printing in pretty mode via `js_layout_generator` we want to preserve newlines between adjacent statements. If two statements are separated by exactly one or two newlines we want to reprint them as separated by the same number of newlines. If statements are separated by more two newlines they should be printed as separated by two newlines.

This was previously being handled by the `list_with_newlines` utility function, which takes in a list of (loc, node) tuples and inserts additional linebreaks between each pair of nodes if their locs are offset by more than one line. However this is causing problems when statements have comments attached, as the loc for a node does not include its attached comments. This means that if a node has a leading comment on the previous line, then an extra newline will always be inserted between it and the previous node, since the two node's locs are separated by more than one line. For example, the following:

```
stmt1;
// L stmt2
stmt2;
```

is being reprinted as:

```
stmt1;

// L stmt2;
stmt2;
```

To fix this we need to know the full location of a statement including all of its leading and trailing comments. This is accomplished with the new `comment_bounds_collector` which uses the AST mapper to find the first leading and last trailing comment attached to a node in the AST or its children.

We then need to pass the "comment_bounds" (the first leading and last trailing comments) for each node to the `list_with_newlines` utility. Then we can expand the loc to include the first leading and last trailing comment, if they exist.

We should also parameterize the separator and the separator's linebreak in `list_with_newlines`, as this same function will be reused in other locations where different separators and line breaking behaviors are required (e.g. between object properties). This change also means that `list_with_newlines` fully inserts all newlines between the input nodes, so we can stop its return value from being wrapped in `join pretty_hardline` at every callsite.

In addition this fixes one other bug where line comments that trail a statement were always having an extra newline inserted after them. This was leading to changes when reprinting, for example:

```
stmt1; // T stmt1
stmt2;
```

was being reprinted as:

```
stmt1; // T stmt1

stmt2;
```

We can fix this in `list_with_newlines` by simply not adding the extra linebreak when the last trailing comment is a line comment, since line comments already end  in a hard newline.

Reviewed By: pieterv

Differential Revision: D21030169

fbshipit-source-id: a0314518f640299fa1a33053418252f89b35108d