Preserve empty lines when printing JSX children
Summary:
When printing in pretty mode we want to preserve formatting in the same way as prettier, so that running prettier after will yield the original file with minimal changes. One such location where we want to  preserve some formatting is between adjacent JSX children.

If there is at least one blank line between adjacent JSX children we should pretty print them as being separated by a single blank line. Luckily JSX children cannot have leading or trailing comments, so it is straightforward to calculate when blank lines should be inserted. We simply need to update the existing newline insertion logic for JSX children to insert a pretty blank line if a JSX child starts on a line that is more than one line past the end of the previous JSX child.

Reviewed By: pieterv

Differential Revision: D21032978

fbshipit-source-id: 0059c4e812137db3a55a89b3acebccac71ed3f42