Groundwork for `flow codemod` command
Summary:
This code is the groundwork for a new `flow codemod` command. This will be
an "umbrella" command that will enable *large-scale codebase refactors*.
This refactors will be later encoded as subcommands whose implementation
will roughly be visitors over the AST.

The various **kinds** of codemods are provided with a *codemod context* that
can be one of two kinds:
* typed context (see Typed_runner): the codemod has access to the typed AST
and the typing contexts. This is the main contribution of this command as it
enables writing codemods that benefit from type inference results.
* untyped context (see Untyped_runner): the context here only includes the
results of parsing (file_sig).

The **processing logic** differs depending on the kind of runner (typed/untyped).
In both versions we first compute a set of input files S that will be transformed.
(See usage for how to provide this set.)
* For typed runners we start a lazy server and mark S as the set to be checked.
Checking might involve a set S' of S. S' is not necessarily a superset of S. (It
may miss files that are not checked by Flow.) We subsequently only apply the
transformation on the intersection of S and S'.
We process files in dependency order since the full contexts and typed ASTs
do not persist in main memory.
Both classic and types-first mode are supported. The transformation is applied
after the merge and check phase, respectively.

* For untyped runners, we try to process roughly the same files that the respective
typed runner would. Since dependencies do not matter in the transformation,
the order of processing is arbitrary.

As **output**, this command, by default, rewrites the target files based on the
transformation. The rewrite is based on the differ-printer, which only replaces
those spans of the program that have been changed. For those parts it uses
the pretty printer to render. The remaining parts are copied over from the original
source. There is also a "dry-run" mode that prints the output on the terminal.

The actual command is not added yet.

Reviewed By: samwgoldman

Differential Revision: D20887580

fbshipit-source-id: f6993a576a275c05299f8fe8dae9b0ecd8537040