[codemod] better reporting for annotate-exports
Summary:
`flow codemod` allows transformations to return a summary value
that is accumulated across the various modules. This value is reported
when all changes are done. Before this change, annotate-exports didn't
use this facility - all reporting was done through logging. This diff
changes this by making this codemod return an summary that includes
statistics about the types that were added and the errors/warnings that
were encountered.

Reviewed By: samwgoldman

Differential Revision: D20908214

fbshipit-source-id: 157194479e90291c75fea883d8cded8e40747ad1