Fix comments being moved during printing of arrow function body
Summary:
We currently have a bug when pretty printing arrow functions with leading comments on the body, and then reformatting the file with prettier.

When an arrow function body has a leading comment, we were previously writing that comment on the same line as the arrow:

```
// Before reprinting
(A) =>
  // $FlowFixMe
  <B />;

// After reprinting
(A) => // $FlowFixMe
  <B />;
```

However when reformatted with prettier, this moves the comment to be attached to the arrow function's argument: [example](https://prettier.io/playground/#N4Igxg9gdgLgprEAuEAKAggSgAQF4B82A9EdgCQBiANhAO4UCWAHgLJwA6U22APAELF8AbhAAaEBAAOMBtADOyUAEMATiroAFVQgUolVWkoCeC8QCMVSsAGs4MAMqSrDKAHNkMFQFc44gBYwALZUAOp+DPByTmBw9joRDABuEUbI4HKmIC5ycCowGpaugUrIAGb6OeIAVnJMfJY2dvZKgXAAMi5wZRW+IDVM9i6uVHAAil4Q8N1UlSBOKjkqaWZKZnBUYnMqLjAhDAAmMH7IABwADOKS6jkhlpJpV3CLiV3iAI4T8AVSuiBKcgBaKBwOD7UGbFRwD4MSEFJRFEpIcozXo5QIMDzeVFDEbjSZdJE9cQwVZ7Q7HJAAJmJlgYVCGAGEIIFimkngBWTZeHIAFVWumRs0SPgAklAwbB7GBttJ0OL7DAjCNpjkAL6qoA).

To fix this we can detect when the body of the arrow function has a leading comment, using the recently created `comment_bounds` utilities. In this case we should print a linebreak in pretty mode after the arrow instead of a space. This will force the comment to be on the line after the arrow, so when reformatting prettier will attach it as leading the body again instead of mistakenly moving it inside the params list.

Reviewed By: pieterv

Differential Revision: D21052611

fbshipit-source-id: 82be588dd55bf80400bc66b0142b669f2130860a